requirements.txt
fastapi
uvicorn
python-multipart
ultralytics
opencv-python-headless
numpy
pillow